feat: 物品悬浮窗优化
* 重构代码，将悬浮窗抽离整理为`ItemPop`
* 现在`ItemSpan`也可以展示悬浮窗 (点击末尾的蓝色信息图标)
* 现在会展示可采集道具的限时信息
* 现在`App`会全局透传当前艾欧泽亚时间`currentET`
