+++ src/tools/item.ts
@@ -126,7 +126,11 @@ export interface ItemInfo {
     placeNameJA: string,
     placeNameEN: string,
     posX: string,
-    posY: string
+    posY: string,
+    timeLimitInfo: {
+      start: string,
+      end: string
+    }[]
   },
   tradeInfo: ItemTradeInfo | undefined
 }
@@ -249,8 +253,18 @@ export const getItemInfo = (item: number | CalculatedItem) => {
           placeNameJA: gatherPlaceData[0],
           placeNameEN: gatherPlaceData[1],
           posX: gatherData.coords.x,
-          posY: gatherData.coords.y
-        }
+          posY: gatherData.coords.y,
+          timeLimitInfo: []
+        };
+        [1,2,3].forEach(i => {
+          if (gatherData?.popTime) {
+            const start = gatherData.popTime?.['start' + i]
+            const end = gatherData.popTime?.['end' + i]
+            if (start && end && start !== '--:--' && end !== '--:--') {
+              itemInfo.gatherInfo.timeLimitInfo.push({ start, end })
+            }
+          }
+        });
       }
     }
   }
